1. More facts about iterations of functions.
2. Started proving some useful lattice properties - e.g. All finite
non-empty lattices are complete, etc. Since I can easily show that an
ascending Kleene chain is a lattice, this would let me show that it is
a complete lattice (provided I can prove that it is finite). Thus we
know the chain has a supremum (which is within the chain) and can use
that to prove Kleene's fixpoint theorem.